% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names_col.R
\name{clean_names_col}
\alias{clean_names_col}
\title{Tidying up column data}
\usage{
clean_names_col(object, case = "snake")
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{case}{Case arguments to be passed to the relevant \pkg{janitor} \code{clean_names} function.}
}
\value{
A modified SingleCellExperiment object.
}
\description{
Tidying up column data
}
\examples{
y <- matrix(rnorm(2000), ncol=20)
se <- SummarizedExperiment(list(counts=y))
se$CellTypes <- sample(LETTERS, ncol(y), replace=TRUE)
se$cell_Types <- sample(LETTERS, ncol(y), replace=TRUE)
clean_names_col(se)
}
\author{
Kevin Wang
}
