% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyexperiment-package.R
\docType{package}
\name{plyexperiment-package}
\alias{plyexperiment}
\alias{plyexperiment-package}
\title{plyexperiment: A fluent interface to SummarizedExperiment}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Stuart Lee \email{lee.s@wehi.edu.au}

Authors:
\itemize{
  \item Michael Lawrence [contributor]
}

Other contributors:
\itemize{
  \item Jonathan Carroll \email{rpkg@jcarroll.com.au} [contributor]
}

}
\keyword{internal}
