% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_split.R
\name{sce_split}
\alias{sce_split}
\title{Splitting a SCE object by a vector}
\usage{
sce_split(sce, f)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{f}{A character/factor vector of the same length as ncol(sce)}
}
\value{
A list of SingleCellExperiment correspond to "f".
}
\description{
Splitting a SCE object by a vector
}
\examples{
y <- matrix(rnorm(2000), ncol=20)
se <- SummarizedExperiment(list(counts=y))
se$CellTypes <- sample(LETTERS, ncol(y), replace=TRUE)
se$cell_Types <- sample(LETTERS, ncol(y), replace=TRUE)
sce_split(se, f = se$CellTypes)
}
\author{
Kevin Wang
}
