% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reducedDim.R
\name{export_reducedDim}
\alias{export_reducedDim}
\title{Exporting the reducedDim object as a tibble}
\usage{
export_reducedDim(sce, type, feature = NULL, assayname = "logcounts")
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{type}{A character in reducedDimNames(sce)}

\item{feature}{A character vector, must be a subset of the rownames of sce}

\item{assayname}{Name of the assay, default to "logcounts"}
}
\value{
A tibble
}
\description{
Exporting the reducedDim object as a tibble
}
\examples{
ncells <- 100
ngenes <- 200
u <- matrix(rpois(ngenes*ncells, 5), ncol=ncells)
colnames(u) = paste0("cell", 1:ncells)
rownames(u) = paste0("gene", 1:ngenes)
v <- log2(u + 1)

colData = S4Vectors::DataFrame(
colrandom = sample(letters, ncells, replace = TRUE),
row.names = colnames(u))

rowData = S4Vectors::DataFrame(
rowrandom = sample(letters, ngenes, replace = TRUE),
row.names = rownames(u))

pca <- matrix(runif(ncells*2), ncells)
tsne <- matrix(rnorm(ncells*2), ncells)

sce <- SingleCellExperiment(assays = list(counts = u, logcounts = v),
colData = colData, rowData = rowData,
reducedDims = SimpleList(PCA = pca, TSNE = tsne))
sce
}
\author{
Kevin Wang
}
